secureBlog/secureblog/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bleach==6.0.0
crispy-bootstrap4==2022.1
Django==4.2
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.2
django-guardian==2.4.0
django-markdownx==4.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-summernote==0.8.20.0
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
webencodings==0.5.1
zipp==3.15.0
